[sdk]: rename MerkleHashBuilder file to merkle
 problem: MerkleHashBuilder has expanded to include additional merkle utils
solution: rename to something more general
