+++ sdk/javascript/src/index.js
@@ -10,10 +10,10 @@ const NemNetwork = require('./nem/Network');
 const NemTransactionFactory = require('./nem/TransactionFactory');
 const NemModels = require('./nem/models');
 const SymbolKeyPair = require('./symbol/KeyPair');
-const SymbolMerkleHashBuilder = require('./symbol/MerkleHashBuilder');
 const SymbolNetwork = require('./symbol/Network');
 const SymbolTransactionFactory = require('./symbol/TransactionFactory');
 const SymbolIdGenerator = require('./symbol/idGenerator');
+const SymbolMerkle = require('./symbol/merkle');
 const SymbolModels = require('./symbol/models');
 const { hexToUint8, uint8ToHex } = require('./utils/converter');
 
@@ -39,10 +39,10 @@ module.exports = {
 	symbol: {
 		...SymbolModels, // must be before Network to promote Address from Network
 		...SymbolKeyPair,
-		...SymbolMerkleHashBuilder,
 		...SymbolNetwork,
 		...SymbolTransactionFactory,
-		...SymbolIdGenerator
+		...SymbolIdGenerator,
+		...SymbolMerkle
 	},
 
 	utils: {
