+++ sdk/javascript/src/symbol/merkle.js
